requirements.txt
argparse
easydict
h5py
matplotlib
numpy<2
open3d
opencv-python
pyyaml
scipy
tensorboardX
timm==0.4.5
tqdm
transforms3d
termcolor
ftfy
regex
ninja
openTSNE
torch-scatter